requirements.txt
ccxt
pandas
requests
python-dotenv
tenacity
pandas-ta
langchain
langchain-google-genai
python-telegram-bot
Flask
Flask-SocketIO
Flask-Limiter
gevent
gevent-websocket